solve remove User bug

+++ src/bin/removeUser.py
@@ -31,26 +31,29 @@ def disconnectDatabase():
 
 try:
     # query existed user to match input account
-    cursor.execute('SELECT userId FROM {} WHERE username = "{}" and password = {};'.format(userTable, username, password))
+    cursor.execute('SELECT userId FROM {} WHERE username = "{}" and password = "{}";'.format(userTable, username, password))
     usersInfo = cursor.fetchall()
 except:
     print("Query fail or incorrect username and password, rollback.")
     db.rollback()
     exit(0)
 
-userId = usersInfo[0][0]
-if userId == 1:
-    print("Can't delete the user: admin")
-    disconnectDatabase()
-    exit(0)
-else:
-    try:
-        cursor.execute('DELETE FROM {} WHERE userId = {};'.format(userTable, userId))
-        cursor.execute('DELETE FROM {} WHERE userId = {};'.format(projectTable, userId))
-        db.commit()
-        print("删除成功")
-    except:
-        print("Delete fail, rollback.")
-        db.rollback()
-    finally:
+if all(usersInfo):
+    print("Unexisted user, please check your account.")
+else:        
+    userId = usersInfo[0][0]
+    if userId == 1:
+        print("Can't delete the user: admin.")
         disconnectDatabase()
+        exit(0)
+    else:
+        try:
+            cursor.execute('DELETE FROM {} WHERE userId = {};'.format(userTable, userId))
+            cursor.execute('DELETE FROM {} WHERE userId = {};'.format(projectTable, userId))
+            db.commit()
+            print("删除成功")
+        except:
+            print("Delete fail, rollback.")
+            db.rollback()
+        finally:
+            disconnectDatabase()
